route3_road/python_client/pyproject.toml
[tool.poetry]
name = "route3_road_client"
version = "0.1.0"
description = "api client for route3"
authors = ["Nico Mandery <nico@nmandery.net>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.38.0"
Shapely = "^1.7.1"
pyarrow = "^4.0.1"
#pandas = "^1.3.0"
#geopandas = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
grpcio-tools = "^1.38.0"
notebook = "^6.4.0"
geopandas = "^0.9.0"
folium = "^0.12.1"
h3 = "^3.7.3"
branca = "^0.4.2"
geojson = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
